* master:
  update version info
  fix code style
  clarify expressions
  cleanups
  if defined(WINSCP)
  if defined(HAVE_OPENSSL)
  cleanup includes
  fix gcc compatibility issues
  update version info
  tweak dlmalloc.cc
  update .vcxproj files
  remove malloc.c
  use dlmalloc.cc
  fix conversion warning
  fix warnings: Zero constant can be replaced with nullptr
  fix warnings: Zero constant can be replaced with nullptr
  fix warnings: Declaration and assignment can be joined
  fix warnings: Declaration and assignment can be joined
  fix code style

# Conflicts:
#	src/base/Classes.cpp

+++ src/base/Classes.cpp
@@ -1195,7 +1195,7 @@ double MilliSecondSpan(const TDateTime & ANow, const TDateTime & AThen)
 
 int64_t MilliSecondsBetween(const TDateTime & ANow, const TDateTime & AThen)
 {
-  int64_t Result = floor(MilliSecondSpan(ANow, AThen));
+  double Result = floor(MilliSecondSpan(ANow, AThen));
   return static_cast<int64_t>(Result);
 }
